Convert `AnyValues` test in `re_dataframe` to snapshot
